symbol: Fix possible multiplication overflows
Multiplication result may overflow 'int' before it is converted to
'unsigned long'.  Practically I don't think it's gonna be a problem
as number of section and program headers are usually small and the
entry size is also very small.

But let's fix it to silence the warnings.

Reported-by: Github Code Scanning Alert

+++ utils/symbol-rawelf.c
@@ -64,7 +64,7 @@ int elf_validate(struct uftrace_elf_data *elf)
 
 	// validate program header offset.
 	size = (long)elf->file_size;
-	offset = ehdr->e_phoff + ehdr->e_phnum * ehdr->e_phentsize;
+	offset = ehdr->e_phoff + (unsigned long)ehdr->e_phnum * ehdr->e_phentsize;
 
 	if (offset > size) {
 		pr_dbg2("Invalid Program Header offset:[%lu], size:[%lu]\n", offset, size);
@@ -72,7 +72,7 @@ int elf_validate(struct uftrace_elf_data *elf)
 	}
 
 	// section header is optional.
-	offset = ehdr->e_shoff + ehdr->e_shnum * ehdr->e_shentsize;
+	offset = ehdr->e_shoff + (unsigned long)ehdr->e_shnum * ehdr->e_shentsize;
 
 	if (offset <= size)
 		elf->has_shdr = true;
